+++ src/backend/EZChat.Master/Database/Repositories/IUserRepository.cs
@@ -6,10 +6,10 @@ namespace EZChat.Master.Database.Repositories
 {
     public interface IUserRepository
     {
-        Task<long> InsertAsync(AppUser user);
+        Task<int> InsertAsync(AppUser user);
         Task UpdateAsync(AppUser user);
         Task DeleteAsync(AppUser user);
-        Task<AppUser> GetByIdAsync(long id);
+        Task<AppUser> GetByIdAsync(int id);
         Task<AppUser> GetByNormalizedUserNameAsync(string normalizedUserName);
     }
 }
